+++ CHANGES.rst
@@ -11,6 +11,8 @@ Release 3.1.0 [2019-XX-XX]
 
 * Enhancements:
 
+  + Add pickle-based ``Request.waitsome()`` and ``Request.testsome()``.
+
   + Support passing Python GPU arrays compliant with the ``__cuda_array_interface__``
     standard (`link`_) to uppercase API calls if mpi4py is built against 
     CUDA-aware MPI implementations. This feature is currently experimental